+++ client/src/game/ui/notes/NoteEdit.vue
@@ -115,6 +115,11 @@ watchEffect(() => {
 // Ensure that defaultAccess is always first
 // and that defaultAccess is provided even if it has no DB value
 const accessLevels = computed(() => {
+    if (note.value === undefined) return [];
+    // For global notes, there is no default access so just return the access levels with a filter
+    if (!note.value.isRoomNote) return note.value.access.filter((a) => a.name !== defaultAccessName);
+    // For local notes, add a default access level and make sure it's first
+    // It's possible that a specific config for the default access level is set, so we need to check for that
     const access = [];
     let defaultAccess = { name: defaultAccessName, can_view: false, can_edit: false };
     for (const a of note.value?.access ?? []) {